moved tests to separate fine
